requirements.txt
django == 4.2.11
djangorestframework == 3.14.0
requests == 2.31.0
pillow == 10.2.0
psycopg2-binary == 2.9.9
celery == 5.3.6
django-celery-beat == 2.6.0
python-dotenv == 1.0.1
python-geohash==0.8.5
django-storages==1.14.2
boto3==1.34.84
channels==4.1.0
daphne==4.1.2